# handlers/auth_handler.go
package handlers

import (
	"encoding/json"
	"net/http"
	"strconv"
	"time"

	"project-MVP/services"
)

type authRequest struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

type registerRequest struct {
	Email    string `json:"email"`
	FullName string `json:"full_name"`
	Role     string `json:"role"`
	Password string `json:"password"`
}

// Register godoc
// @Summary Регистрация пользователя
// @Description Создает новый аккаунт пользователя в системе Scifi. Доступные роли: admin, user, guest.
// @Tags auth
// @Accept json
// @Produce json
// @Param request body registerRequest true "Данные регистрации"
// @Success 200 {object} map[string]interface{} "Данные созданного пользователя"
// @Failure 400 {string} string "Ошибка валидации или дубликат пользователя"
// @Failure 500 {string} string "Внутренняя ошибка сервера"
// @Router /auth/register [post]
func Register(w http.ResponseWriter, r *http.Request) {
	var req registerRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	user, err := services.RegisterUser(req.Email, req.FullName, req.Role, req.Password)
	if err != nil {
		if err == services.ErrInvalidRole || err == services.ErrDuplicateUser {
			http.Error(w, err.Error(), http.StatusBadRequest)
			return
		}
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]any{
		"id":        user.Id,
		"email":     user.Email,
		"full_name": user.FullName,
		"role":      user.Role,
	})
}

// Login godoc
// @Summary Авторизация
// @Description Аутентификация пользователя по Email и паролю. После успешного входа устанавливается сессионная кука 'session'.
// @Tags auth
// @Accept json
// @Produce json
// @Param request body authRequest true "Данные входа"
// @Success 200 {object} map[string]int "ID авторизованного пользователя"
// @Failure 401 {string} string "Неверный логин или пароль"
// @Failure 500 {string} string "Внутренняя ошибка сервера"
// @Router /auth/login [post]
func Login(w http.ResponseWriter, r *http.Request) {
	var req authRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	userID, err := services.AuthenticateUser(req.Email, req.Password)
	if err != nil {
		if err == services.ErrInvalidCredentials {
			http.Error(w, err.Error(), http.StatusUnauthorized)
			return
		}
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	cookie := &http.Cookie{
		Name:     "session",
		Value:    strconv.Itoa(userID),
		Path:     "/",
		HttpOnly: true,
		SameSite: http.SameSiteLaxMode,
		Expires:  time.Now().Add(24 * time.Hour),
	}
	http.SetCookie(w, cookie)

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]any{
		"user_id": userID,
	})
}

// Logout godoc
// @Summary Выход из системы
// @Description Удаляет сессионную куку и перенаправляет на страницу логина
// @Tags auth
// @Success 303 {string} string "Перенаправление на /login"
// @Router /logout [get]
func Logout(w http.ResponseWriter, r *http.Request) {
	cookie := &http.Cookie{
		Name:     "session",
		Value:    "",
		Path:     "/",
		HttpOnly: true,
		MaxAge:   -1,
		Expires:  time.Unix(0, 0),
	}
	http.SetCookie(w, cookie)
	http.Redirect(w, r, "/login", http.StatusSeeOther)
}